+++ enterpark-ticket-domain/src/main/kotlin/com/example/enterparkticket/domain/performance/model/repository/PerformanceRepository.kt
@@ -1,4 +1,4 @@
-package com.example.enterparkticket.domain.performance.repository
+package com.example.enterparkticket.domain.performance.model.repository
 
 import com.example.enterparkticket.domain.performance.model.Performance
 import org.springframework.data.jpa.repository.JpaRepository
