+++ enterpark-ticket-domain/src/main/kotlin/com/example/enterparkticket/domain/performance/query/PerformanceQueryService.kt
@@ -2,7 +2,7 @@ package com.example.enterparkticket.domain.performance.query
 
 import com.example.enterparkticket.domain.performance.exception.NotFoundPerformanceException
 import com.example.enterparkticket.domain.performance.model.Performance
-import com.example.enterparkticket.domain.performance.repository.PerformanceRepository
+import com.example.enterparkticket.domain.performance.model.repository.PerformanceRepository
 import org.springframework.stereotype.Service
 import org.springframework.transaction.annotation.Transactional
