+++ enterpark-ticket-domain/src/main/kotlin/com/example/enterparkticket/domain/seat/command/SeatCommandService.kt
@@ -1,6 +1,6 @@
 package com.example.enterparkticket.domain.seat.command
 
-import com.example.enterparkticket.domain.seat.repository.SeatRepository
+import com.example.enterparkticket.domain.seat.model.repository.SeatRepository
 import org.springframework.stereotype.Service
 import org.springframework.transaction.annotation.Transactional
