+++ enterpark-ticket-domain/src/main/kotlin/com/example/enterparkticket/domain/user/command/UserCommandService.kt
@@ -1,7 +1,7 @@
 package com.example.enterparkticket.domain.user.command
 
 import com.example.enterparkticket.domain.user.model.User
-import com.example.enterparkticket.domain.user.repository.UserRepository
+import com.example.enterparkticket.domain.user.model.repository.UserRepository
 import org.springframework.stereotype.Service
 import org.springframework.transaction.annotation.Transactional
