+++ enterpark-ticket-domain/src/main/kotlin/com/example/enterparkticket/domain/user/model/repository/UserRepository.kt
@@ -1,4 +1,4 @@
-package com.example.enterparkticket.domain.user.repository
+package com.example.enterparkticket.domain.user.model.repository
 
 import com.example.enterparkticket.domain.user.model.OAuthInfo
 import com.example.enterparkticket.domain.user.model.User
